test(os): Add context, sleep2, sync2 test suites — 30 tests all passing
New tests:
- context: R4-R11 register preservation across yields (with a
  clobber task writing different values), stack isolation between
  tasks using canary patterns
- sleep2: repeated sleep_ms(10) x50 accumulation, sleep_ms(1)
  minimum guarantee, equal-priority fairness (35% minimum share),
  high-priority preemption on wake
- sync2: counting semaphore (give 5, take 5), multiple waiters
  on one semaphore (3 blocked, give 3, all wake)

Also widened wait_queue.waiters to uint32_t (was uint16_t) to
support up to 32 tasks in the test suite.

Results: 30 tests, 30 passing, 0 flaky (10/10 clean runs).

+++ os/kernel/sched.h
@@ -37,7 +37,7 @@ enum task_state {
  * Like Zephyr's _wait_q_t / Linux's wait_queue_head_t.
  */
 struct wait_queue {
-    uint16_t waiters;  /* bitmask of task IDs waiting */
+    uint32_t waiters;  /* bitmask of task IDs waiting */
 };
 
 #define WAIT_QUEUE_INIT { .waiters = 0 }
